+++ build_runner/lib/build_runner.dart
@@ -2,5 +2,5 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/entrypoint/daemon.dart' show assetServerPort;
+export 'src/daemon/constants.dart' show assetServerPort;
 export 'src/entrypoint/run.dart' show run;
